refactor: modularize article JS into likes and comments handlers
